[Ref] Splitting Dummy backend into classes, pt. 3
